Fix review comments: move waitForData to the test utils file
